+++ crates/teloxide-core/src/adaptors/cache_me.rs
@@ -217,7 +217,8 @@ where
         set_game_score_inline,
         get_game_high_scores,
         approve_chat_join_request,
-        decline_chat_join_request
+        decline_chat_join_request,
+        get_available_gifts
         => f, fty
     }
 }