+++ crates/teloxide-core/src/requests/requester.rs
@@ -1315,6 +1315,11 @@ pub trait Requester {
     fn get_game_high_scores<T>(&self, user_id: UserId, target: T) -> Self::GetGameHighScores
     where
         T: Into<TargetMessage>;
+
+    type GetAvailableGifts: Request<Payload = GetAvailableGifts, Err = Self::Err>;
+
+    /// For Telegram documentation see [`GetAvailableGifts`].
+    fn get_available_gifts(&self) -> Self::GetAvailableGifts;
     // END BLOCK requester_methods
 }
 
@@ -1464,7 +1469,8 @@ macro_rules! forward_all {
             set_game_score_inline,
             get_game_high_scores,
             approve_chat_join_request,
-            decline_chat_join_request
+            decline_chat_join_request,
+            get_available_gifts
             => $body, $ty
         }
     };